dapogrill/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-phone-field==1.8.1
Flask==2.0.2
future==0.18.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.8.1
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
Paperboy==1.0.1
Pillow==8.1.0
psycopg2==2.9.3
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
-e git+https://github.com/gbozee/django-paystack.git@24df1c9ca209031e0ec940ff8552d45ca6b55421#egg=pypaystack
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.10.0.2
urllib3==1.26.3
Werkzeug==2.0.2
whitenoise==6.0.0
wrapt==1.12.1
zipp==3.6.0
